added basic renderer (this is how it is shown on the molecule musings blog) added memory arena, linear allocators and other memory helpers (this has been taken from the molecule musings blog too and will be extended soon) moved the timing stuff into functions and a struct that contains all necessary information added a allocate cube method to the primitive interfaces switched the existing rendering of the spaceship to use the new renderer (primitives are still missing for this) imgui is being called directly from the deferred renderer class (this will be moved to the new renderer structure, too once I work on editor stuff)

+++ Dawn/Dawn/src/Dawn/Core/Type.h
@@ -12,9 +12,9 @@ namespace Dawn
 		friend class TypeMemberCollection;
 	public:
 		std::string Name;
-		u128 Size;
+		u64 Size;
 
-		Type(std::string InName, u128 InSize, TypeInitMethod InInit)
+		Type(std::string InName, u64 InSize, TypeInitMethod InInit)
 			: Name(InName), Size(InSize) 
 		{
 			if(InInit != nullptr)
@@ -33,7 +33,7 @@ namespace Dawn
 		}
 
 		
-		void AddMember(std::string InName, u128 InSize) 
+		void AddMember(std::string InName, u64 InSize) 
 		{
 			TypeMember member = { InName, InSize };
 			Members.emplace(std::make_pair(InName, member));
@@ -43,7 +43,7 @@ namespace Dawn
 		struct TypeMember
 		{
 			std::string Name;
-			u128 Offset;
+			u64 Offset;
 		};
 
 		std::map<std::string, TypeMember> Members;
